requirements.txt
Python-IO
sqlalchemy
regex
Pillow
easyocr==1.7.1
opencv-python-headless
Pillow==10.0.1
streamlit==1.27.2
pandas==2.1.1
numpy==1.26.0
altair==4.2.2
attrs==22.2.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.0.4
colorama==0.4.6
decorator==5.1.1
entrypoints==0.4
fire==0.5.0
geographiclib==2.0
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.30
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mdurl==0.1.2
packaging==23.0
plotly==5.13.0
protobuf==3.20.3
pyarrow==11.0.0
pybboxes==0.1.6
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
PyMySQL==1.0.2
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
rich==13.3.1
sahi==0.11.11
semver==2.13.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
streamlit-image-comparison==0.0.3
tenacity==8.2.1
termcolor==2.2.0
terminaltables==3.1.10
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
watchdog==2.2.1
zipp==3.13.0
